+++ src/feedback/GEAR/stellar_wind.h
@@ -40,8 +40,8 @@ void stellar_wind_read_yields_array(
 
 void stellar_wind_clean(struct stellar_wind *sw);
 
-double stellar_wind_get_ejected_energy(const struct stellar_wind *sw, double log_m, float log_z);
+double stellar_wind_get_ejected_energy(const struct stellar_wind *sw, float log_m, float log_z);
 
-double stellar_wind_get_ejected_energy_IMF(const struct stellar_wind *sw, double log_m, float log_z);
+double stellar_wind_get_ejected_energy_IMF(const struct stellar_wind *sw, float log_m, float log_z);
 
 #endif
